Twitter-released IO Campaign Metadata

Metadata describing data archives that Twitter believes are part of potentially state-backed information operations campaigns.

io_campaign_metadata

Format

An object of class data.table (inherits from data.frame) with 22 rows and 9 columns.

Source

https://transparency.twitter.com/en/information-operations.html

Details

Twitter updates the data archives available on a regular basis, so these may be out of date. You can help by filing issues at https://github.com/knapply/tweetio/issues.

The following information is from Twitter:

"Note that by downloading these files, you are accepting the Twitter Developer Agreement and Policy."

"These datasets are of a size that a degree of capability for large dataset analysis is required, we hope to support broad analysis by making a public version of these datasets (with some account-specific information hashed) available. You can download the datasets below. No content has been redacted. Specialist researchers can request access to an unhashed version of these datasets, which will be governed by a data use agreement that will include provisions to ensure the data is used within appropriate legal and ethical parameters."

"These datasets include all public, nondeleted Tweets and media (e.g., images and videos) from accounts we believe are connected to state-backed information operations. Tweets deleted by these users prior to their suspension (which are not included in these datasets) generally comprise less than 1% of their overall activity. Note that not all of the accounts we identified as connected to these campaigns actively Tweeted, so the number of accounts represented in the datasets may be less than the total number of accounts listed here."

Examples

io_campaign_metadata[, .(campaign_name, date_released, entities)]

#>             campaign_name date_released                   entities
#>  1:                   ira    2018-10-01   Internet Research Agency
#>  2:               iranian    2018-10-01                       Iran
#>  3:         iran_201901_1    2019-01-01                       Iran
#>  4:   bangladesh_201901_1    2019-01-01                 Bangladesh
#>  5:       russia_201901_1    2019-01-01                     Russia
#>  6:    venezuela_201901_1    2019-01-01                  Venezuela
#>  7:    venezuela_201901_2    2019-01-01                  Venezuela
#>  8:    catalonia_201906_1    2019-06-01                  Catalonia
#>  9:         iran_201906_1    2019-06-01                       Iran
#> 10:         iran_201906_2    2019-06-01                       Iran
#> 11:         iran_201906_3    2019-06-01                       Iran
#> 12:       russia_201906_1    2019-06-01                     Russia
#> 13:    venezuela_201906_1    2019-06-01                  Venezuela
#> 14:        china_082019_1    2019-08-01                      China
#> 15:        china_082019_2    2019-08-01                      China
#> 16:        china_082019_3    2019-09-01                      China
#> 17: saudi_arabia_082019_1    2019-09-01               Saudi Arabia
#> 18:      ecuador_082019_1    2019-09-01                    Ecuador
#> 19:          uae_082019_1    2019-09-01       United Arab Emirates
#> 20:        spain_082019_1    2019-09-01                      Spain
#> 21:    egypt_uae_082019_1    2019-09-01 United Arab Emirates,Egypt
#> 22:   saudi_arabia_112019    2019-12-01               Saudi Arabia
#>             campaign_name date_released                   entities