Convert Various Objects to proto_net

A proto_net is a list containing two data frames named edges and nodes.

as_proto_net(tweet_df, target_class = c("user", "hashtag", "url",
  "media"), all_status_data = FALSE, all_user_data = FALSE,
  as_tibble = FALSE, ...)

Arguments

tweet_df
A data frame of tweets, as obtained by read_tweets() or one of {rtweet}'s collection functions, e.g. rtweet::search_tweets().

target_class
character(1L), Default: "user". The class of nodes to use as the second half of each dyad (target/to/tail). See Details.

all_status_data
logical(1L), Default: FALSE. Whether to attach all relevant status data to the edges data frame, which can then be used as edge attributes for downstream tasks.

all_user_data
logical(1L), Default: FALSE. Whether to attach all relevant user data to the nodes data frame, which can then be uses as node attributes for downstream tasks..

as_tibble
Whether a tibble::tibble() should be returned. Ignored if the {tibble} package is not installed.

...
Arguments passed to or from other methods.

Details

In a proto_net, users are always to source/from/head side of dyads. target_class defaults to "user", which creates edges where users on both sides of dyads.

However users can also share edges with "hashtag"s, "url"s, or "media", so those values are also valid to provide to target_class to create two-mode/bipartite proto_nets.

The edges of a proto_net represent the statuses that form each tie, and status-specific columns are attached to the edges.

Casing

Twitter hashtags are not case-sensitive, so if target_class is "hashtag" they will be cast to lower-case so they can represent the same node in downstream tasks.

URLs (after the domain) can be case-sensitive, so they are left as-is.

If you decide to cast URLs to lower-case yourself, exercise caution with Twitter's media URLs, as they contain a case-sensitive hash.

all_status_data and all_user_data default to FALSE as they can be performance bottlenecks for large data sets, but they provide a way of building richly decorated networks with the maximum amount of attribute data embedded in the graph structure.

Examples

path_to_tweet_file <- example_tweet_file()

tweet_df <- read_tweets(path_to_tweet_file)

tweet_df %>%
  as_proto_net(as_tibble = TRUE)

#> $edges
#> # A tibble: 42,191 x 4
#>    from                to                  status_id           relation
#>    <chr>               <chr>               <chr>               <chr>   
#>  1 807195027403259904  938662230052052992  1178007813257187328 retweet 
#>  2 247059655           609647782           1178007813232025601 retweet 
#>  3 822602638231535617  701750402786717696  1178007813236244481 retweet 
#>  4 1104415177603133444 1108885040862121985 1178007813236252672 retweet 
#>  5 359294371           1034795669645447168 1178007813248806912 retweet 
#>  6 194250838           340309688           1178007813257388032 retweet 
#>  7 1160132819114287105 3251156738          1178007813257224192 retweet 
#>  8 1064053347164930048 2430693075          1178007813244608513 retweet 
#>  9 1056020503993888768 1017892758063702016 1178007813244604417 retweet 
#> 10 1103702847378604032 414928075           1178007813253206017 retweet 
#> # ... with 42,181 more rows
#> 
#> $nodes
#> # A tibble: 37,519 x 1
#>    name               
#>    <chr>              
#>  1 807195027403259904 
#>  2 247059655          
#>  3 822602638231535617 
#>  4 1104415177603133444
#>  5 359294371          
#>  6 194250838          
#>  7 1160132819114287105
#>  8 1064053347164930048
#>  9 1056020503993888768
#> 10 1103702847378604032
#> # ... with 37,509 more rows
#> 
#> attr(,"class")
#> [1] "proto_net"
#> attr(,"target_class")
#> [1] "user"

tweet_df %>%
  as_proto_net(target_class = "hashtag", as_tibble = TRUE)

#> $edges
#> # A tibble: 7,174 x 4
#>    from                to                status_id           relation
#>    <chr>               <chr>             <chr>               <chr>   
#>  1 359294371           got7_kconth2019   1178007813248806912 uses    
#>  2 933502505115377665  investigatebiden  1178007813257338882 uses    
#>  3 981315512956268545  <U+C154><U+B204>              1178007813236375555 uses    
#>  4 767835057620918272  therookie         1178007813265805312 uses    
#>  5 781414129           <U+0623><U+062D><U+062F>_<U+0627><U+0644><U+0645><U+0633><U+0627><U+0631><U+062D><U+0647>      1178007813257388034 uses    
#>  6 781414129           <U+0627><U+0644><U+064A><U+0648><U+0645>_<U+0627><U+0644><U+0648><U+0637><U+0646><U+064A>89    1178007813257388034 uses    
#>  7 781414129           <U+0647><U+0645><U+0647>_<U+062D><U+062A><U+0649>_<U+0627><U+0644><U+0642><U+0645><U+0647>     1178007813257388034 uses    
#>  8 1868079872          wicklowecotrail   1178007813257334785 uses    
#>  9 1177957168869298177 naamkarannke3saal 1178007817422163968 uses    
#> 10 2162617110          watch             1178007817443086336 uses    
#> # ... with 7,164 more rows
#> 
#> $nodes
#> # A tibble: 7,522 x 1
#>    name               
#>    <chr>              
#>  1 359294371          
#>  2 933502505115377665 
#>  3 981315512956268545 
#>  4 767835057620918272 
#>  5 781414129          
#>  6 1868079872         
#>  7 1177957168869298177
#>  8 2162617110         
#>  9 1169328651222913024
#> 10 3112061905         
#> # ... with 7,512 more rows
#> 
#> attr(,"class")
#> [1] "proto_net"
#> attr(,"target_class")
#> [1] "hashtag"

tweet_df %>%
  as_proto_net(target_class = "url", as_tibble = TRUE)

#> $edges
#> # A tibble: 2,382 x 4
#>    from         to                                        status_id     relation
#>    <chr>        <chr>                                     <chr>         <chr>   
#>  1 93045527748~ https://twitter.com/valsnudes/status/117~ 117800781325~ uses    
#>  2 11110051225~ https://twitter.com/LVPibai/status/11770~ 117800781324~ uses    
#>  3 76783505762~ https://www.distractify.com/p/does-offic~ 117800781326~ uses    
#>  4 86349961067~ https://twitter.com/sacramentokings/stat~ 117800781324~ uses    
#>  5 3859220422   http://blog.pianetadonna.it/mybeautyblog~ 117800781746~ uses    
#>  6 10116681785~ https://headlines.yahoo.co.jp/hl?a=20190~ 117800781743~ uses    
#>  7 2933680894   https://www.30millionsdamis.fr/conseils/~ 117800781746~ uses    
#>  8 11738773892~ https://ift.tt/2mHkZx1                    117800781745~ uses    
#>  9 856475624    https://twitter.com/sarichuga_many/statu~ 117800782162~ uses    
#> 10 401300296    https://www.instagram.com/p/B29ty1WI8Hq/~ 117800782164~ uses    
#> # ... with 2,372 more rows
#> 
#> $nodes
#> # A tibble: 4,442 x 1
#>    name               
#>    <chr>              
#>  1 930455277488562176 
#>  2 1111005122543472640
#>  3 767835057620918272 
#>  4 863499610676973569 
#>  5 3859220422         
#>  6 1011668178504540160
#>  7 2933680894         
#>  8 1173877389241323521
#>  9 856475624          
#> 10 401300296          
#> # ... with 4,432 more rows
#> 
#> attr(,"class")
#> [1] "proto_net"
#> attr(,"target_class")
#> [1] "url"

tweet_df %>%
  as_proto_net(target_class = "media", as_tibble = TRUE)

#> $edges
#> # A tibble: 5,446 x 4
#>    from         to                                        status_id     relation
#>    <chr>        <chr>                                     <chr>         <chr>   
#>  1 247059655    http://pbs.twimg.com/ext_tw_video_thumb/~ 117800781323~ uses    
#>  2 359294371    http://pbs.twimg.com/ext_tw_video_thumb/~ 117800781324~ uses    
#>  3 10560205039~ http://pbs.twimg.com/ext_tw_video_thumb/~ 117800781324~ uses    
#>  4 93350250511~ http://pbs.twimg.com/ext_tw_video_thumb/~ 117800781325~ uses    
#>  5 84458315     http://pbs.twimg.com/tweet_video_thumb/E~ 117800781324~ uses    
#>  6 2259128973   http://pbs.twimg.com/media/EFkWlRwUwAEyl~ 117800781324~ uses    
#>  7 199369785    http://pbs.twimg.com/media/EFkf2xcX0AAZt~ 117800781322~ uses    
#>  8 97761929941~ http://pbs.twimg.com/media/EFkf0zdWsAYvl~ 117800781326~ uses    
#>  9 781414129    http://pbs.twimg.com/media/EFkf2O5X0AE4I~ 117800781325~ uses    
#> 10 781414129    http://pbs.twimg.com/media/EFkf2O8W4AEa5~ 117800781325~ uses    
#> # ... with 5,436 more rows
#> 
#> $nodes
#> # A tibble: 10,126 x 1
#>    name               
#>    <chr>              
#>  1 247059655          
#>  2 359294371          
#>  3 1056020503993888768
#>  4 933502505115377665 
#>  5 84458315           
#>  6 2259128973         
#>  7 199369785          
#>  8 977619299413168129 
#>  9 781414129          
#> 10 821026391072325633 
#> # ... with 10,116 more rows
#> 
#> attr(,"class")
#> [1] "proto_net"
#> attr(,"target_class")
#> [1] "media"

tweet_df %>%
  as_proto_net(all_status_data = TRUE, all_user_data = TRUE, as_tibble = TRUE)

#> $edges
#> # A tibble: 42,191 x 20
#>    from  to    status_id relation created_at          text  status_url source
#>    <chr> <chr> <chr>     <chr>    <dttm>              <chr> <chr>      <chr> 
#>  1 1127~ 1108~ 11780103~ retweet  2019-09-28 18:15:22 RT @~ https://t~ Twitt~
#>  2 1127~ 1108~ 11780103~ mentions 2019-09-28 18:15:22 RT @~ https://t~ Twitt~
#>  3 9857~ 1563~ 11780103~ retweet  2019-09-28 18:15:22 RT @~ https://t~ Twitt~
#>  4 9857~ 1563~ 11780103~ mentions 2019-09-28 18:15:22 RT @~ https://t~ Twitt~
#>  5 1099~ 1453~ 11780103~ reply_to 2019-09-28 18:15:22 @has~ https://t~ Twitt~
#>  6 1099~ 1453~ 11780103~ mentions 2019-09-28 18:15:22 @has~ https://t~ Twitt~
#>  7 9645~ 1071~ 11780103~ retweet  2019-09-28 18:15:22 RT @~ https://t~ Twitt~
#>  8 9645~ 1071~ 11780103~ mentions 2019-09-28 18:15:22 RT @~ https://t~ Twitt~
#>  9 1129~ 1139~ 11780103~ reply_to 2019-09-28 18:15:22 @age~ https://t~ Twitt~
#> 10 1129~ 1139~ 11780103~ mentions 2019-09-28 18:15:22 @age~ https://t~ Twitt~
#> # ... with 42,181 more rows, and 12 more variables: is_quote <lgl>,
#> #   is_retweeted <lgl>, media_url <list>, media_type <list>, place_url <chr>,
#> #   place_name <chr>, place_full_name <chr>, place_type <chr>, country <chr>,
#> #   country_code <chr>, bbox_coords <list>, status_type <chr>
#> 
#> $nodes
#> # A tibble: 37,519 x 20
#>    name  timestamp_ms        name.y screen_name location description url  
#>    <chr> <dttm>              <chr>  <chr>       <chr>    <chr>       <chr>
#>  1 1000~ 2019-09-28 18:14:16 ©<U+FE0F>0<U+FE0F><U+20E3>STA  _costinhaaa Portugal CERCI BOY ~ NA   
#>  2 1000~ 2019-09-28 18:10:12 "Mari~ mariana_rgr NA       NA          NA   
#>  3 1000~ 2019-09-28 18:07:15 flore~ florence__t NA       NA          NA   
#>  4 1000~ 2019-09-28 18:08:38 "Lela~ Daniela12x  Guimarã~ "ig: x.dan~ NA   
#>  5 1000~ 2019-09-28 18:13:19 "\U00~ Mcfbb_      Viseu, ~ "@fcporto ~ NA   
#>  6 1000~ 2019-09-28 18:10:14 "Bapt~ BaptCorre1  Brest, ~ @Portos130~ NA   
#>  7 1000~ 2019-09-28 18:13:30 Alexa~ AlexLopezM~ Colombia Senador de~ NA   
#>  8 1000~ 2019-09-28 18:06:39 Léa    Lea_Lefebv~ "bsm "   on s’était~ NA   
#>  9 1000~ 2019-09-28 18:12:52 "R \U~ ritaddvaz   Portale~ NA          http~
#> 10 1000~ 2019-09-28 18:09:47 Subro~ subrotosir~ Kanchra~ Be Yoursel~ NA   
#> # ... with 37,509 more rows, and 13 more variables: protected <lgl>,
#> #   followers_count <int>, friends_count <int>, listed_count <int>,
#> #   statuses_count <int>, favourites_count <int>, account_created_at <dttm>,
#> #   verified <lgl>, profile_url <chr>, account_lang <chr>,
#> #   profile_banner_url <chr>, profile_image_url <chr>, bbox_coords <list>
#> 
#> attr(,"class")
#> [1] "proto_net"
#> attr(,"target_class")
#> [1] "user"