Convert Various Objects to proto_nets

A proto_net is a list containing two data frames named edges and nodes.

as_proto_net(
  tweet_df,
  target_class = c("user", "hashtag", "url", "media"),
  all_status_data = FALSE,
  all_user_data = FALSE,
  as_tibble = tweetio_as_tibble(),
  ...
)

# S3 method for data.frame
as_proto_net(
  tweet_df,
  target_class = c("user", "hashtag", "url", "media"),
  all_status_data = FALSE,
  all_user_data = FALSE,
  as_tibble = tweetio_as_tibble(),
  ...
)

# S3 method for data.table
as_proto_net(
  tweet_df,
  target_class = c("user", "hashtag", "url", "media"),
  all_status_data = FALSE,
  all_user_data = FALSE,
  as_tibble = tweetio_as_tibble(),
  ...
)

Arguments

tweet_df
A data frame of tweets, as obtained by read_tweets() or one of {rtweet}'s collection functions, e.g. rtweet::search_tweets().

target_class
character(1L), Default: "user". The class of nodes to use as the second half of each dyad (target/to/head). See Details.

all_status_data
logical(1L), Default: FALSE. Whether to attach all relevant status data to the edges data frame, which can then be used as edge attributes for downstream tasks.

all_user_data
logical(1L), Default: FALSE. Whether to attach all relevant user data to the nodes data frame, which can then be uses as node attributes for downstream tasks.

as_tibble
<logical>, Default: tweetio_as_tibble(). Whether a tibble::tibble() should be returned. Ignored if the {tibble} package is not installed.

...
Arguments passed to or from other methods.

Details

In a proto_net, users are always to source/from/tail side of dyads. target_class defaults to "user", which creates edges where users on both sides of dyads.

However users can also share edges with "hashtag"s, "url"s, or "media", so those values are also valid to provide to target_class to create two-mode/bipartite proto_nets.

The edges of a proto_net represent the statuses that form each tie, and status-specific columns are attached to the edges.

Casing

Twitter hashtags are not case-sensitive, so if target_class is "hashtag" they will be cast to lower-case so they can represent the same node in downstream tasks.

URLs (after the domain) can be case-sensitive, so they are left as-is.

If you decide to cast URLs to lower-case yourself, exercise caution with Twitter's media URLs, as they contain a case-sensitive hash.

all_status_data and all_user_data default to FALSE as they can be performance bottlenecks for large data sets, but they provide a way of building richly decorated networks with the maximum amount of attribute data embedded in the graph structure.

Examples

path_to_tweet_file <- example_tweet_file()

tweet_df <- read_tweets(path_to_tweet_file)

tweet_df %>%
  as_proto_net(as_tibble = TRUE)

#> $edges
#> # A tibble: 1,234 x 5
#>    from            to              status_id        created_at          relation
#>    <chr>           <chr>           <chr>            <dttm>              <chr>   
#>  1 194250838       340309688       117800781325738… 2019-09-29 02:05:23 retweet 
#>  2 82545948782161… 966825602       117800781742637… 2019-09-29 02:05:24 retweet 
#>  3 11162285596163… 3153145782      117800783003444… 2019-09-29 02:05:27 retweet 
#>  4 4374655520      4167284315      117800783000508… 2019-09-29 02:05:27 retweet 
#>  5 11728856250680… 11128778918413… 117800783841898… 2019-09-29 02:05:29 retweet 
#>  6 11324745945839… 1217220278      117800784259214… 2019-09-29 02:05:30 retweet 
#>  7 20737729        39334221        117800784679069… 2019-09-29 02:05:31 retweet 
#>  8 2198859787      3096758526      117800784678228… 2019-09-29 02:05:31 retweet 
#>  9 3237877098      84788518732527… 117800785100173… 2019-09-29 02:05:32 retweet 
#> 10 86821900935067… 1962155772      117800785940313… 2019-09-29 02:05:34 retweet 
#> # … with 1,224 more rows
#> 
#> $nodes
#> # A tibble: 1,228 x 1
#>    name               
#>    <chr>              
#>  1 194250838          
#>  2 825459487821619201 
#>  3 1116228559616397312
#>  4 4374655520         
#>  5 1172885625068036102
#>  6 1132474594583928832
#>  7 20737729           
#>  8 2198859787         
#>  9 3237877098         
#> 10 868219009350676481 
#> # … with 1,218 more rows
#> 
#> attr(,"class")
#> [1] "proto_net"
#> attr(,"target_class")
#> [1] "user"

tweet_df %>%
  as_proto_net(target_class = "hashtag", as_tibble = TRUE)

#> $edges
#> # A tibble: 267 x 5
#>    from            to               status_id       created_at          relation
#>    <chr>           <chr>            <chr>           <dttm>              <chr>   
#>  1 218889555       sooners          11780078174600… 2019-09-28 19:05:24 uses    
#>  2 93732206        conejopride      11780078677705… 2019-09-28 19:05:36 uses    
#>  3 93732206        inclusionishapp… 11780078677705… 2019-09-28 19:05:36 uses    
#>  4 93732206        ireadbannedbooks 11780078677705… 2019-09-28 19:05:36 uses    
#>  5 11655549854063… ป๋อจ้าน            11780078887253… 2019-09-28 19:05:41 uses    
#>  6 1270106738      theboss          11780078887422… 2019-09-28 19:05:41 uses    
#>  7 1270106738      jefa             11780078887422… 2019-09-28 19:05:41 uses    
#>  8 1270106738      nahreptop250     11780078887422… 2019-09-28 19:05:41 uses    
#>  9 1270106738      nahrep           11780078887422… 2019-09-28 19:05:41 uses    
#> 10 1270106738      olivaresandmoli… 11780078887422… 2019-09-28 19:05:41 uses    
#> # … with 257 more rows
#> 
#> $nodes
#> # A tibble: 366 x 1
#>    name               
#>    <chr>              
#>  1 218889555          
#>  2 93732206           
#>  3 1165554985406304257
#>  4 1270106738         
#>  5 1065721261744107521
#>  6 1177166063710035968
#>  7 979945309370572800 
#>  8 110467298          
#>  9 1177725296306540545
#> 10 16429082           
#> # … with 356 more rows
#> 
#> attr(,"class")
#> [1] "proto_net"
#> attr(,"target_class")
#> [1] "hashtag"

tweet_df %>%
  as_proto_net(target_class = "url", as_tibble = TRUE)

#> $edges
#> # A tibble: 162 x 5
#>    from      to                         status_id   created_at          relation
#>    <chr>     <chr>                      <chr>       <dttm>              <chr>   
#>  1 401300296 https://www.instagram.com… 1178007821… 2019-09-28 19:05:25 uses    
#>  2 43746555… https://twitter.com/richi… 1178007830… 2019-09-28 19:05:27 uses    
#>  3 21988597… https://twitter.com/_trap… 1178007846… 2019-09-28 19:05:31 uses    
#>  4 32378770… http://M.Tech              1178007851… 2019-09-28 19:05:32 uses    
#>  5 93732206  https://twitter.com/i/web… 1178007867… 2019-09-28 19:05:36 uses    
#>  6 13627947… https://twitter.com/i/web… 1178007876… 2019-09-28 19:05:38 uses    
#>  7 22047720… https://www.instagram.com… 1178007884… 2019-09-28 19:05:40 uses    
#>  8 12701067… https://www.instagram.com… 1178007888… 2019-09-28 19:05:41 uses    
#>  9 10657212… https://twitter.com/i/web… 1178007909… 2019-09-28 19:05:46 uses    
#> 10 11771660… https://twitter.com/parit… 1178007926… 2019-09-28 19:05:50 uses    
#> # … with 152 more rows
#> 
#> $nodes
#> # A tibble: 316 x 1
#>    name               
#>    <chr>              
#>  1 401300296          
#>  2 4374655520         
#>  3 2198859787         
#>  4 3237877098         
#>  5 93732206           
#>  6 1362794708         
#>  7 2204772030         
#>  8 1270106738         
#>  9 1065721261744107521
#> 10 1177166063710035968
#> # … with 306 more rows
#> 
#> attr(,"class")
#> [1] "proto_net"
#> attr(,"target_class")
#> [1] "url"

tweet_df %>%
  as_proto_net(target_class = "media", as_tibble = TRUE)

#> $edges
#> # A tibble: 137 x 5
#>    from      to                         status_id   created_at          relation
#>    <chr>     <chr>                      <chr>       <dttm>              <chr>   
#>  1 11728856… http://pbs.twimg.com/medi… 1178007838… 2019-09-28 19:05:29 uses    
#>  2 11324745… http://pbs.twimg.com/ext_… 1178007842… 2019-09-28 19:05:30 uses    
#>  3 93732206  http://pbs.twimg.com/ext_… 1178007867… 2019-09-28 19:05:36 uses    
#>  4 11214555… http://pbs.twimg.com/medi… 1178007884… 2019-09-28 19:05:40 uses    
#>  5 10657212… http://pbs.twimg.com/medi… 1178007909… 2019-09-28 19:05:46 uses    
#>  6 10657212… http://pbs.twimg.com/medi… 1178007909… 2019-09-28 19:05:46 uses    
#>  7 10657212… http://pbs.twimg.com/medi… 1178007909… 2019-09-28 19:05:46 uses    
#>  8 10429113… http://pbs.twimg.com/medi… 1178007943… 2019-09-28 19:05:54 uses    
#>  9 110467298 http://pbs.twimg.com/medi… 1178007947… 2019-09-28 19:05:55 uses    
#> 10 98041717… http://pbs.twimg.com/twee… 1178007964… 2019-09-28 19:05:59 uses    
#> # … with 127 more rows
#> 
#> $nodes
#> # A tibble: 253 x 1
#>    name               
#>    <chr>              
#>  1 1172885625068036102
#>  2 1132474594583928832
#>  3 93732206           
#>  4 1121455512283832320
#>  5 1065721261744107521
#>  6 1042911305202454528
#>  7 110467298          
#>  8 980417171531685888 
#>  9 2328141540         
#> 10 834583812977881090 
#> # … with 243 more rows
#> 
#> attr(,"class")
#> [1] "proto_net"
#> attr(,"target_class")
#> [1] "media"

tweet_df %>%
  as_proto_net(all_status_data = TRUE, all_user_data = TRUE, as_tibble = TRUE)

#> $edges
#> # A tibble: 1,234 x 20
#>    from  to    status_id created_at          relation text  i.created_at       
#>    <chr> <chr> <chr>     <dttm>              <chr>    <chr> <dttm>             
#>  1 8627… 8453… 11780103… 2019-09-29 02:15:22 retweet  "RT … 2019-09-29 02:15:22
#>  2 8627… 8453… 11780103… 2019-09-29 02:15:22 mentions "RT … 2019-09-29 02:15:22
#>  3 8627… 1579… 11780103… 2019-09-29 02:15:22 mentions "RT … 2019-09-29 02:15:22
#>  4 8627… 1508… 11780103… 2019-09-29 02:15:22 mentions "RT … 2019-09-29 02:15:22
#>  5 8627… 5943… 11780103… 2019-09-29 02:15:22 mentions "RT … 2019-09-29 02:15:22
#>  6 3688… 8698… 11780103… 2019-09-29 02:15:20 retweet  "RT … 2019-09-29 02:15:20
#>  7 3688… 8698… 11780103… 2019-09-29 02:15:20 mentions "RT … 2019-09-29 02:15:20
#>  8 3688… 7661… 11780103… 2019-09-29 02:15:20 mentions "RT … 2019-09-29 02:15:20
#>  9 3688… 1876… 11780103… 2019-09-29 02:15:20 mentions "RT … 2019-09-29 02:15:20
#> 10 3688… 5882… 11780103… 2019-09-29 02:15:20 mentions "RT … 2019-09-29 02:15:20
#> # … with 1,224 more rows, and 13 more variables: source <chr>, is_quote <lgl>,
#> #   is_retweeted <lgl>, media_url <list>, media_type <list>, place_url <chr>,
#> #   place_name <chr>, place_full_name <chr>, place_type <chr>, country <chr>,
#> #   country_code <chr>, bbox_coords <list>, status_type <chr>
#> 
#> $nodes
#> # A tibble: 1,228 x 19
#>    name  timestamp_ms        name.y screen_name location description url  
#>    <chr> <dttm>              <chr>  <chr>       <chr>    <chr>       <chr>
#>  1 1001… 2019-09-29 02:15:00 ブリアナ.… yoongi_Far… 新宿     "⎷ = 私は秘密の… NA   
#>  2 1001… 2019-09-29 02:13:04 Local… ornella944… Bronx, … "Snap:orne… NA   
#>  3 1001… 2019-09-29 02:08:35 RAWAN… rawanjehad… Jeddah   "Respirato… http…
#>  4 1001… 2019-09-29 02:12:13 Shash… sks_rk      Kolkata… "Right to … NA   
#>  5 1003… 2019-09-29 02:06:45 J🥀    cherrydoze  Garland… "3/22🥝"    http…
#>  6 1003… 2019-09-29 02:13:16 Side03 Side031     Space F… "Twitter d… NA   
#>  7 1003… 2019-09-29 02:09:58 J.O.K… AlexisAbat… NA       "Money Lif… NA   
#>  8 1003… 2019-09-29 02:07:53 ti_be… t_benjy     Bourges… "Bad Vibes… NA   
#>  9 1003… 2019-09-29 02:11:28 A A R… ManojManuu_ India    "@alluarju… NA   
#> 10 1003… 2019-09-29 02:06:26 백혀_엑소… i8Cg9ETNlz… Taipei … "@B_hundre… http…
#> # … with 1,218 more rows, and 12 more variables: protected <lgl>,
#> #   followers_count <int>, friends_count <int>, listed_count <int>,
#> #   statuses_count <int>, favourites_count <int>, account_created_at <dttm>,
#> #   verified <lgl>, account_lang <chr>, profile_banner_url <chr>,
#> #   profile_image_url <chr>, bbox_coords <list>
#> 
#> attr(,"class")
#> [1] "proto_net"
#> attr(,"target_class")
#> [1] "user"